task 0

+++ 0x01-python_async_function/0-basic_async_syntax.py
@@ -15,7 +15,7 @@ import asyncio
 import random
 
 
-async def wait_random(max_delay: int = 10) -> float:
+async def wait_random(max_delay: float = 10) -> float:
     """
     Asynchronous function that waits for a
     random delay and returns the delay.
